% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lsurvROC.R
\name{plot.lsurvROC}
\alias{plot.lsurvROC}
\title{Plot method for lsurvROC objects}
\usage{
\method{plot}{lsurvROC}(x, ROC = FALSE, ...)
}
\arguments{
\item{ROC}{logic value}

\item{model_results}{A lsurvROC object.}
}
\description{
Custom plot for lsurvROC class.
}
