% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lsurvROC.R
\name{plot_lsurvROC}
\alias{plot_lsurvROC}
\title{plot_lsurvROC}
\usage{
plot_lsurvROC(
  model,
  my.newdat,
  tau,
  basis,
  tol = 1000,
  add = FALSE,
  col = "black",
  lty = 1,
  main = NULL
)
}
\arguments{
\item{model}{a model result from lsurvROC( )}

\item{my.newdat}{a data.frame used for calculating the conditional ROC curve}

\item{tau}{a vector of tau values used in the model}

\item{basis}{basis function used in the model}

\item{tol}{a constant used for removing unconverged results}

\item{add}{if TURE, add the ROC curve on an existing plot}

\item{col}{color of the curve}

\item{lty}{line type}

\item{main}{plot title}
}
\description{
It returns the conditional ROC curve and the area under curve (AUC) value.
}
\examples{
  model.results <- lsurvROC(dat.long = example_data$data.long, 
    dat.short = example_data$data.short,
    cutoff.type.basis = "FP",
    sens.type.basis = "FP", 
    covariate1 = c("Z", "Zcont"), 
    covariate2 = c("Z", "Zcont"), 
    tau = seq(0.1, 0.9, 0.05), 
    time.window = 1, 
    nResap = 50, 
    show_plots = FALSE
    )
 
 ROC.results <- plot_lsurvROC(model = model.results$models, 
   my.newdat = data.frame(vtime = 0.5, Z = 1, Zcont = 0.25), 
    tau = seq(0.1, 0.9, 0.05),
    basis = "FP", 
    tol = 1e3, 
    add = FALSE,
    col = "black", 
    lty = 1)
    
 ROC.results$AUC  
 
}
