% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lsurvROC.R
\name{plot.lsurvROC}
\alias{plot.lsurvROC}
\title{Plot method for lsurvROC objects}
\usage{
\method{plot}{lsurvROC}(x, ROC = FALSE, ...)
}
\arguments{
\item{x}{a lsurvROC object}

\item{ROC}{a logical value that indicates whether plot the ROC curve}
}
\description{
Custom plot for lsurvROC class.
}
