% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lsurvROC.R
\name{plot.lsurvROC}
\alias{plot.lsurvROC}
\title{Plot method for lsurvROC objects}
\usage{
\method{plot}{lsurvROC}(x, ROC = FALSE, newdata = NULL, ROC.tau = NULL, ...)
}
\arguments{
\item{x}{A lsurvROC object}

\item{ROC}{logic value}

\item{newdata}{logic valu}

\item{ROC.tau}{logic valu}
}
\description{
Custom plot for lsurvROC class.
}
