% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lsurvROC.R
\name{plot.lsurvROC}
\alias{plot.lsurvROC}
\title{Plot method for lsurvROC objects}
\usage{
\method{plot}{lsurvROC}(x, tol = 1000, ROC = FALSE, ...)
}
\arguments{
\item{x}{A lsurvROC object}

\item{tol}{logic value}

\item{ROC}{logic value}
}
\description{
Custom plot for lsurvROC class.
}
